Add reference station and rover support for ZED-FP9 module
The ZED-FP9 module does not have "REF" or "ROV" in their product 
description, which would result in only the HpPosRecProduct component 
being added. The ZED-FP9 can act as both a base station and a rover.

Also updates the NavRELPOSENED message to publish under a new topic for 
the firmware 9 version, so they do not conflict.