Add support for publishing and consuming RTCM corrections
This enables the ublox driver to parse RTCM messages from the serial 
data, publish those corrections on a topic, as well as consume 
corrections on a topic and send them to the device over. This does not 
do any parsing of the corrections data, it simply broadcasts the entire 
message.

Changes to enable this are:
-Modify the callback handler processing the serial buffer to check for 
both UBLOX and RTCM messages, this necessitated some new functionality 
with ther serialization process to expose the size of the ublox messages 
being read
-Add a rtcm callback to the callback handler, which is provided by a 
higher level
-Add a publisher to the HpgRefProduct class to publish the RTCM 
corrections as a UInt8MultiArray message
-Add a subscriber to the HpgRovProduct for RTCM corrections in 
UInt8MultiArray messages, which will send those corrections to the 
device 
when received